The solution partially copies the logic of the built-in HashMap, with the exception of threshold, which is initialized immediately.
Overridden put(K key, V value) method from the MyMap<K, V> interface.
Since the newly created MyHashMap == null, we must initialize it in the resize() method. Also, resize() increases the size of the MyHashMap array when the number of values reaches a critical point.
After that, we add a new node.

Overridden method from interface MyMap<K, V>.
It returns value by key and expects that the bucket may contain a List. Otherwise returns null.

Overridden getSize() method from MyMap<K, V> interface.
The method returns the size;

Added class Node.
The hash variable is excluded from it, as well as the equals and hashCode methods due to the specifics of the task and implementation.

Added private void putValue(int hash, K key, V value).
It expects adding Node: with key null; duplicate; and generates a list in the bucket if it was already occupied by another Node.

Added private void method resize().
It initializes an empty hashTable when the first element is added.
It also doubles the size of the array if necessary, and also calls the transfer() method to transfer nodes from the old table to the new one.

Added private void transfer(Node<K, V>[] oldHashTable) method.
Its task is to through all existing elements in the old table and use putValue() to determine them to buckets in the new table.

Added private int hash(K key) method.
With it, we determine in which bucket will be the new Node.