The logic for solving this problem was as follows: first, I needed to create a method that would allow me to change the size of the hash table and prevent collisions.
resize():
 The first step is to invoke the "resize" method, which enables changing the size of the hash table.
Assign the new length to the variable "newLength".
Preserve references to all elements by storing the old array, "table", in the variable "oldArray".
Create a new array, "table", with the updated length using the syntax "table = new Node[newLength];".
Set the size of the hash table to zero with the statement "size = 0;", as all elements will be added again.
Iterate through all the elements of the old array using a loop.
Set the current node to the first node in the list located at a certain index, using the assignment "current = kvNode;".
Utilize a while() loop to iterate through each node in the list.
Call the "put()" method, which adds the current key and value to the hash table. This operation overwrites the value if the key already exists or adds a new key-value pair if the key doesn't exist yet.
Move to the next node in the list by assigning "current = current.next;".
getIndex():

This method is responsible for computing the index in the hash table based on the given key.
Using the ternary operator (?), the method checks if the key is null. If true, it returns the value 0.
If the key is not null, the "hashCode" method for the key is called, returning its hash code. The "Math.abs()" function is applied to obtain the absolute value of the hash code.

getNode():
This method is responsible for retrieving a node from the hash table based on the key.